fixing warnings
git-svn-id: svn://sidvind.com/project/BTH/NPLDP/libcap_utils/trunk@2023 8459abcf-8237-4014-9f1e-057f68ddaf13

+++ caputils_int.h
@@ -73,7 +73,7 @@ int stream_tcp_init(struct stream* st, const char* address, int port);
 long stream_ethernet_open(struct stream** stptr, const struct ether_addr* address, const char* iface);
 long stream_ethernet_create(struct stream** stptr, const struct ether_addr* address, const char* iface, const char* mpid, const char* comment);
 
-int stream_file_open(struct stream** stptr, const char* filename);
+long stream_file_open(struct stream** stptr, const char* filename);
 
 /**
  * @param fp Optional
